Rename fake server internals to servers
