+++ src/servers/whatsapp-wire/binary-node.ts
